Major code refactor
I've essentially stopped hardcoding stuff like my social links and instead I'll put it all in config.json and read from that config file when I need to. This makes it easier to setup a custom website for local software development purposes, and it makes it easier for me to update those social links if need be.